refactor; address comments
