+++ linode_api4/groups/__init__.py
@@ -10,7 +10,7 @@ from .lke import *
 from .longview import *
 from .networking import *
 from .nodebalancer import *
-from .obj import *
+from .object_storage import *
 from .profile import *
 from .region import *
 from .support import *
