+++ test/objects/object_storage_test.py
@@ -1,7 +1,11 @@
 from datetime import datetime
 from test.base import ClientBaseCase
 
-from linode_api4.objects import ObjectStorageACL, ObjectStorageBucket
+from linode_api4.objects import (
+    ObjectStorageACL,
+    ObjectStorageBucket,
+    ObjectStorageCluster,
+)
 
 
 class ObjectStorageTest(ClientBaseCase):
@@ -89,3 +93,29 @@ class ObjectStorageTest(ClientBaseCase):
                 },
             )
             self.assertEqual(m.call_url, bucket_access_update_url)
+
+    def test_buckets_in_cluster(self):
+        """
+        Test that Object Storage Buckets in a specified cluster can be reterived
+        """
+        buckets_in_cluster_url = "/object-storage/buckets/us-east-1"
+        with self.mock_get(buckets_in_cluster_url) as m:
+            cluster = ObjectStorageCluster(self.client, "us-east-1")
+            buckets = cluster.buckets_in_cluster()
+            self.assertIsNotNone(buckets)
+            bucket = buckets[0]
+
+            self.assertEqual(m.call_url, buckets_in_cluster_url)
+            self.assertEqual(bucket.cluster, "us-east-1")
+            self.assertEqual(
+                bucket.created,
+                datetime(
+                    year=2019, month=1, day=1, hour=1, minute=23, second=45
+                ),
+            )
+            self.assertEqual(
+                bucket.hostname, "example-bucket.us-east-1.linodeobjects.com"
+            )
+            self.assertEqual(bucket.label, "example-bucket")
+            self.assertEqual(bucket.objects, 4)
+            self.assertEqual(bucket.size, 188318981)
